added more components to the dashboard
