+++ LMS.Blazor/LMS.Blazor.Client/Components/Courses/CourseSummaryCard.razor
@@ -5,7 +5,7 @@
 @inherits Microsoft.AspNetCore.Components.ComponentBase
 
 <div class="col-4">
-    <div class="card p-3">
+    <div class="card p-3 mb-3">
         <div class="d-flex justify-content-between align-items-center mb-1">
             <BookOpenIcon CssClass="text-primary" />
             <span class="badge rounded-pill @(CourseIsActive ? "bg-success-subtle text-success" : "bg-danger-subtle text-danger")">
@@ -13,7 +13,7 @@
             </span>
         </div>
 
-        <h5><strong>@Course.Name</strong></h5>
+        <h5 class="fw-bold">@Course.Name</h5>
         <p>@Course.Description</p>
 
         <div class="pb-3">
@@ -29,18 +29,20 @@
             <span>@Course.ModulesCount modules</span>
         </div>
 
-        <hr />
+        <hr class="text-secondary" />
 
         <ProgressBar StartDate="@Course.StartDate" EndDate="@Course.EndDate" />
 
-        <button class="btn btn-primary btn-sm"
-                @onclick="() => OnShowParticipants.InvokeAsync(Course.Id)">
-            Show participants
-        </button>
-        <button class="btn btn-outline-secondary btn-sm ms-2"
-                @onclick="() => OnShowModules.InvokeAsync(Course.Id)">
-            Show modules
-        </button>
+        <div class="d-flex justify-content-between align-items-center mb-1">
+            <button class="btn btn-primary btn-sm"
+                    @onclick="() => OnShowParticipants.InvokeAsync(Course.Id)">
+                Show participants
+            </button>
+            <button class="btn btn-outline-secondary btn-sm ms-2"
+                    @onclick="() => OnShowModules.InvokeAsync(Course.Id)">
+                Show modules
+            </button>
+        </div>
     </div>
 </div>
