+++ LMS.Blazor/LMS.Blazor.Client/Pages/Teacher/Dashboard.razor
@@ -1,5 +1,4 @@
 ﻿@page "/teacher/dashboard"
-@inherits DashboardBase
 @using LMS.Blazor.Client.Components
 @using LMS.Blazor.Client.Components.Courses
 @using LMS.Blazor.Client.Components.Icons
@@ -9,10 +8,10 @@
 @using LMS.Shared.DTOs.Module
 @attribute [Authorize(Roles = "Teacher")]
 
+
 <PageTitle>Dashboard – Lexicon LMS</PageTitle>
 
 <ParticipantsModal Title="Participants" Visible="@showParticipantsModal" VisibleChanged="@(v => showParticipantsModal = v)" Items="participants" />
-
 <ModulesModal Title="Modules" Visible="@showModulesModal" VisibleChanged="@(v => showModulesModal = v)" Items="modulesList" />
 
 <div class="d-flex align-items-center justify-content-between mb-4 flex-wrap gap-2">
@@ -22,7 +21,14 @@
                                     HiddenText="Create course" />
 </div>
 
-<CourseSummarySearch SearchText="@searchText" SearchTextChanged="@(v => searchText = v)" ActiveFilter="@activeFilter" ActiveFilterChanged="@(v => activeFilter = v)" OnSearchInput="OnSearchDebounced" OnActiveChanged="OnActiveChanged" OnSearchClicked="@( () => LoadCourses(1) )" />
+<CourseSummarySearch 
+    SearchText="@searchText" 
+    SearchTextChanged="@(v => searchText = v)" 
+    ActiveFilter="@activeFilter" 
+    ActiveFilterChanged="@(v => activeFilter = v)" 
+    OnSearchInput="OnSearchDebounced" 
+    OnActiveChanged="OnActiveChanged" 
+    OnSearchClicked="@( () => LoadCourses(1) )" />
 
 @if (_showCreateForm)
 {
@@ -53,5 +59,5 @@
 </div>
 
 <CourseSummaryPagination CurrentPage="@currentPage" PageSize="@pageSize" TotalCount="@totalCount" OnPageChanged="@(p => LoadCourses(p))" />
-
+<TeacherToDo />
