+++ LMS.Blazor/LMS.Blazor.Client/Pages/Teacher/TeacherToDo.razor
@@ -1,0 +1,17 @@
+﻿<h4 class="fw-bold my-4">To Do</h4>
+
+<div class="card px-3 mb-4">
+    <div class="d-flex justify-content-between align-items-center">
+        <ul>
+           <li>
+                <h6 class="fw-bold">New assignment from Maria Larsson</h6>
+                <p class="text-muted">Submitted: 2026-03-30</p>
+           </li>
+       </ul>
+
+       <button class="btn btn-outline-secondary btn-sm ms-2">Handle</button>
+    </div>
+</div>
+
+
+<h4 class="fw-bold mb-0">Courses</h4>
